It works! (see description)
DB is basically done. Need to complete modification and deletion of credential sets, but the reading and writing to the DB is functional. Some formatting needs to be done, UserInputManager needs to be completed and cleaned up. Comments need to be fixed, and don't mind the unnecessary super()'s, they're part of the project requirements...